requirements.txt
flask
flask-sqlachemy
flask-migrate
flask-bcrypt
flask-wtf
flask-login
flask-admin
flask-limiter
flask-caching
flask-socketio
wtforms

python-dotenv
gunicorn
email-validator
huggingface_hub
matplotlib
seaborn
pandas
sqlachemy
PIL
werkzueg